Add bbs-auth-entity and property files
