--- pd-apps/bbs-auth/bbs-auth-entity/src/main/java/com/metaverseapp/bbs/authority/entity/common/OptLog.java
@@ -1,4 +1,0 @@
-package com.metaverseapp.bbs.authority.entity.common;
-
-public class OptLog {
-}
