+++ pd-apps/bbs-auth/bbs-auth-entity/src/main/java/net/metaverseapp/bbs/authority/entity/common/OptLog.java
@@ -1,0 +1,4 @@
+package net.metaverseapp.bbs.authority.entity.common;
+
+public class OptLog {
+}
