+++ pd-apps/bbs-auth/bbs-auth-entity/src/main/java/net/metaverseapp/bbs/authority/enumeration/common/LogType.java
@@ -1,0 +1,4 @@
+package net.metaverseapp.bbs.authority.enumeration.common;
+
+public class LogType {
+}
