+++ pd-apps/bbs-auth/bbs-auth-server/src/main/java/net/metaverseapp/bbs/authority/biz/dao/common/LoginLogMapper.java
@@ -1,0 +1,4 @@
+package net.metaverseapp.bbs.authority.biz.dao.common;
+
+public class LoginLogMapper {
+}
