+++ pd-apps/bbs-auth/bbs-auth-server/src/main/java/net/metaverseapp/bbs/authority/biz/service/auth/UserService.java
@@ -1,0 +1,4 @@
+package net.metaverseapp.bbs.authority.biz.service.auth;
+
+public interface UserService {
+}
