+++ pd-apps/bbs-auth/bbs-auth-server/src/main/java/net/metaverseapp/bbs/authority/config/ExceptionConfiguration.java
@@ -1,0 +1,4 @@
+package net.metaverseapp.bbs.authority.config;
+
+public class ExceptionConfiguration {
+}
