+++ pd-apps/bbs-auth/bbs-auth-server/src/main/java/net/metaverseapp/bbs/authority/controller/auth/UserController.java
@@ -1,0 +1,4 @@
+package net.metaverseapp.bbs.authority.controller.auth;
+
+public class UserController {
+}
