+++ pd-apps/bbs-auth/bbs-auth-server/src/main/java/net/metaverseapp/bbs/authority/controller/common/OptLogController.java
@@ -1,0 +1,4 @@
+package net.metaverseapp.bbs.authority.controller.common;
+
+public class OptLogController {
+}
