+++ pd-apps/bbs-auth/bbs-auth-server/src/main/resources/j2cache/caffeine.properties
@@ -1,0 +1,9 @@
+#########################################
+# Caffeine configuration
+# \u6682\u65F6\u6CA1\u7528
+# [name] = size, xxxx[s|m|h|d]
+#########################################
+default=2000, 2h
+captcha=1000, 5m
+resource=2000, 2h
+user_resource=3000, 2h
