+++ pd-apps/bbs-auth/bbs-auth-server/src/main/resources/logback-spring.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<configuration>
+    <!--引入pd-tools-log模块中提供的日志的基础配置-->
+    <include resource="com/itheima/pinda/log/logback/pinda-defaults.xml"/>
+
+    <springProfile name="test,docker,prod">
+        <logger name="com.itheima.pinda.authority.controller" additivity="true"
+                level="${log.level.controller}">
+            <appender-ref ref="ASYNC_CONTROLLER_APPENDER"/>
+        </logger>
+        <logger name="com.itheima.pinda.authority.biz.service" additivity="true"
+                level="${log.level.service}">
+            <appender-ref ref="ASYNC_SERVICE_APPENDER"/>
+        </logger>
+        <logger name="com.itheima.pinda.authority.biz.dao" additivity="false"
+                level="${log.level.dao}">
+            <appender-ref ref="ASYNC_DAO_APPENDER"/>
+        </logger>
+    </springProfile>
+
+    <springProfile name="dev">
+        <logger name="com.itheima.pinda.authority.controller" additivity="true"
+                level="${log.level.controller}">
+            <appender-ref ref="CONTROLLER_APPENDER"/>
+        </logger>
+        <logger name="com.itheima.pinda.authority.biz.service" additivity="true"
+                level="${log.level.service}">
+            <appender-ref ref="SERVICE_APPENDER"/>
+        </logger>
+    </springProfile>
+</configuration>
