+++ pd-apps/bbs-auth/bbs-auth-server/src/main/resources/spy.properties
@@ -1,0 +1,10 @@
+module.log=com.p6spy.engine.logging.P6LogFactory,com.p6spy.engine.outage.P6OutageFactory
+logMessageFormat=com.baomidou.mybatisplus.extension.p6spy.P6SpyLogger
+appender=com.baomidou.mybatisplus.extension.p6spy.StdoutLogger
+deregisterdrivers=true
+useprefix=true
+excludecategories=info,debug,result,commit,resultset
+dateformat=yyyy-MM-dd HH:mm:ss
+driverlist=com.mysql.cj.jdbc.Driver
+outagedetection=true
+outagedetectioninterval=2
